requirements.txt
alpaca-trade-api==3.2.0
pandas>=2.2.2
numpy>=1.26.4
python-dotenv>=1.0.1
requests>=2.32.3
types-requests==2.28.11
apscheduler>=3.10.4
beautifulsoup4>=4.12.3
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.9
httpx>=0.27.0
prometheus-client>=0.20.0
redis>=5.0.4
duckdb>=0.10.2
respx>=0.21.1
pytest-asyncio>=0.23.5
structlog>=24.1.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-jaeger>=1.21.0
scikit-learn>=1.4.2
aiohttp>=3.9.5
aioresponses>=0.7.4
wikipedia>=1.4.0
tqdm>=4.66.4
unidecode>=1.3.8
praw>=7.7.1
yfinance>=0.2.40
torch>=2.3.0
transformers>=4.41.1
sentencepiece>=0.2.0
mypy>=1.8.0

pydantic>=2.5.0
pyarrow>=14.0.2
sentry-sdk>=2.0.0
pydantic-settings>=2.0.3
Deprecated>=1.2.13
PyYAML>=6.0
